feat: create auth provider
The auth provider creates a simpler interface to interact with the login
/ logout functionality of the application. Additionally all the context
providers will be aggregated in the AppProviders component.
