+++ index.html
@@ -5,13 +5,12 @@
     <link rel="icon" type="image/svg+xml" href="/src/styles/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-
     <title>backslash</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script src="https://accounts.google.com/gsi/client" async defer></script>
+    <script src="https://accounts.google.com/gsi/client"></script>
     <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
