+++ src/auth/index.ts
@@ -1,2 +1,2 @@
 export * from "./sign-in-with-google";
-export * from "./use-auth";
+export * from "./auth";
